+++ src/main/java/com/SmartBank/repository/CustomerRepository.java
@@ -1,15 +1,14 @@
 package com.SmartBank.repository;
 
 
-import com.SmartBank.model.Customer;
+import com.SmartBank.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
     boolean existsByEmail(String email);
 
     boolean existsByPhone(String phone);
